Move services into a dedicated namespace
