+++ OpenKh.Tools.Kh2MapCollisionEditor/Services/CollisionMapDrawHandler.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Xe.Drawing;
 
-namespace OpenKh.Tools.Kh2MapCollisionEditor
+namespace OpenKh.Tools.Kh2MapCollisionEditor.Services
 {
     public class CollisionMapDrawHandler
     {
